Fix assumptions that sizeof(int) == sizeof(long) which broke on amd64
git-svn-id: svn://red.pomac.com/rage@183 55c13d69-e7e2-0310-b639-a4a306287c42

+++ trunk/plugins/python/python.c
@@ -1421,7 +1421,7 @@ Module_rage_get_prefs(PyObject *self, PyObject *args)
 			break;
 		case 2:
 		case 3: /* Is this even 64bit safe?! */
-			res = PyInt_FromLong((int)info);
+			res = PyInt_FromLong((long)info);
 			break;
 		default:
 			PyErr_Format(PyExc_RuntimeError,
